refactor(data_operations): consolidate aggregation-family bases (closes #246)
The aggregation, window_aggregation and scalar_aggregate base classes
carried byte-for-byte identical aggregation-type machinery
(`_extract_aggregation_type`, `get_aggregation_type`,
`_validate_string_match`, the `return_data_type_rule` skeleton) plus a
duplicated agg-type constant table (a fourth partial copy lived in
frame_aggregate).

Introduce a shared `aggregation_base.py` holding the canonical
`AGGREGATION_TYPES` table and `AggregationFeatureGroupBase`, mirroring the
`arithmetic_base.py` consolidation from #214. The three families now
subclass it; per-family differences stay as small overrides:

- window keeps its own first/last descriptions (spread over the canonical
  table) for the "(requires order_by)" wording.
- scalar binds its 13-key subset as the class `AGGREGATION_TYPES` and
  overrides `_COUNTING_AGG_TYPES` to {"count"} (it has no nunique).
- frame_aggregate is a light touch: it keeps its own matching and
  `return_data_type_rule` (placeholder PREFIX_PATTERN, `_extract_params`),
  and only derives its `_AGGREGATION_TYPES` set from the shared canonical
  table so the fourth copy is gone.

No behavior change. The counting set is exposed as the `_COUNTING_AGG_TYPES`
class hook so the `return_data_type_rule` skeleton can be shared without
changing which agg types declare INT64. tox green; framework-support-matrix
drift check unchanged.
